Add files via upload
So this was some version of me trying to do the x value "addition" and "multiplication" and then I'm starting on the polynomials. I figure if we had then everyone else's would be easier to build. Some of it was just crappy to try to understand but actually doing the methods has probably really helped me to understand what it's doing at least math wise. 
I used bitset to represent the bytes/words because it was just generally easier for me to rap my head around it. The only problem is that the coef multiplication is kinda weird with the <128> and I've attempted it a few times but this version I just erased it so that it will compile for the time being. In main you can see some test cases. Modulo was also extremely weird to get. 
I know it's not exactly what I was suppose to do this week, but I will try to just do this the rest of the day since I actually have some time today. I will upload when complete.