Add files via upload
Changed all the bitsets that were parameters or returned by functions. There still are a few to make the modulo stuff wayyy easier and plus I think it's unnecessary for at least now. This IS NOT BEEN TESTED in its new form, but was working well before. I think I understand and did what the paper was explaining.

However, I think this will be extremely useful for other functions that still need to be built/redone. Please make use of these bc I think they are important to understand. Remember index 0 is least significant. vector<vector<char>> is tricky bc I feel like maybe I'm mixing up columns/ rows but I don't think so actually. the inside vector distinguishes the rows and the outside deals with the columns, as that's how mixcolumns() works.